start work on redis ACL security for device registration
The goal here is going to be to have the devices flashed with redis
authentication information that can only blpop from the specific
`ob:r` registrar index.

Subsequent commands with be required to have authenticated using the id
requested, which will have been locked down by the registar using an
additional ACL entry.

This commit also adds the `src/beetle-pio-tls-tester` project which can
be used to verify certificate selection.